backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
pymongo==4.14.1
pydantic==2.11.7
python-multipart==0.0.20
PyPDF2==3.0.1
pdfplumber==0.11.7
pandas==2.3.2
openpyxl==3.1.5
reportlab==4.4.4
requests==2.32.5
python-dotenv==1.2.2
openai-whisper
torch
ffmpeg-python
